Rewrite device detection logic and synchronize device hot-plugging
Before on quick hotplugging (detaching and reattaching) of a device,
it could happen that the detached event was triggered after the attached
event. As a result features could be unregistered falsely. New devices
are now added to a queue as long as a device with the same DeviceUid is
still present. Only after the previous device closes, the new
device will be opened from that queue.