Add abstraction for feature activities (layout and device handling)
So far all feature based activities have been copies of the same
boilerplate for layout and device handling. This is now taken care of
within FeatureActivity. A new kind of viewable feature can now be
created with ease. FeatureActivity also allows to define multiple
views and statistics as FeatureFragments with very little to no code
duplication.