Add multi-touch drag and drop user positioning to overlay bubble
The overlay bubble had a static position before. This changeset enables
the user to drag and drop the bubble to any position on screen. The
volume controls have been updated accordingly to work along all screen
borders. To make testing easier, the overlay now has a demo version in
the overlay settings, that can be activated without a device.